Restrict use of Vorbis comment functions and simplify
vcedit_write needs most of the functionality of vcedit_open to skip over
the header packets so their use is restricted as follows:

 1. The reader state passed to vcedit_write must be from vcedit_open
    with the input stream at the right position (with Vorbis comments
    updated as required).

 2. Once vcedit_write has used the reader state, it cannot be reused by
    another call to vcedit_write.

Given 2, vcedit_write clears the reader state, which allows it to be
reused with vcedit_open.  The restrictions on usage are documented.

vcedit_write is also updated to properly support NULL for the error
argument (although not currently called with such an error value).